+++ src/features/Guild/actions.js
@@ -1,4 +1,6 @@
-import { get } from 'axios';
+// @flow
+
+import axios from 'axios';
 import config from 'config';
 import { browserHistory } from 'react-router';
 
@@ -6,12 +8,12 @@ export const FETCHING_GUILD = 'FETCHING_GUILD';
 export const SELECT_GUILD = 'SELECT_GUILD';
 export const FETCH_GUILD_RESULT = 'FETCH_GUILD_RESULT';
 
-const fetchingGuild = (fetching) => ({
+const fetchingGuild = (fetching: boolean) => ({
   type: FETCHING_GUILD,
   payload: fetching,
 });
 
-export const selectGuild = (name) => ({
+export const selectGuild = (name: string) => ({
   type: SELECT_GUILD,
   payload: name,
 });
@@ -24,10 +26,10 @@ const fetchGuildResult = (name, data) => ({
   },
 });
 
-export const fetchGuild = (name) => (dispatch) => {
+export const fetchGuild = (name: string) => (dispatch: Dispatch) => {
   dispatch(fetchingGuild(true));
 
-  return get(`${config.api.endpoint}guilds/${name}`)
+  return axios.get(`${config.api.endpoint}guilds/${name}`)
     .then((response) => {
       dispatch(fetchGuildResult(name, response.data));
       dispatch(fetchingGuild(false));