+++ src/features/Guild/index.js
@@ -2,9 +2,17 @@
 
 import { Component } from 'react';
 import { connect } from 'react-redux';
+import T from 'i18n-react';
 
+import type { Guild as GuildType } from 'flowTypes';
+
+import SvgIcon from 'common/components/Icon/Svg';
 import Content from 'common/layouts/Content';
 import ContentCardList from 'common/components/ContentCardList';
+import TooltipTrigger from 'common/components/TooltipTrigger';
+import Overview from './components/Overview';
+
+import styles from './styles.less';
 
 import {
   selectGuild,
@@ -18,11 +26,7 @@ import { selector } from './guilds.reducer';
 })
 export default class Guild extends Component {
   props: {
-    guild: {
-      tag: string,
-      characters?: [],
-      users?: [],
-    },
+    guild?: GuildType,
     routeParams: {
       guildName: string,
     },
@@ -39,19 +43,37 @@ export default class Guild extends Component {
   }
 
   render () {
-    const { guild = { tag: '...', characters: undefined, users: undefined }, routeParams: { guildName } } = this.props;
+    const { guild, routeParams: { guildName } } = this.props;
     const encodedGuildName = encodeURI(guildName);
+
+    const claimed = guild && guild.claimed;
+    const claimedData = {
+      logo: claimed ? 'done' : 'error-outline',
+      message: claimed ? T.translate('guilds.claimed') : T.translate('guilds.unclaimed'),
+    };
 
     return (
       <Content
-        title={`${guildName} [${guild.tag}]`}
+        title={`${guildName} [${(guild && guild.tag) || '...'}]`}
+        cardExtra={(
+          <TooltipTrigger data={claimedData.message}>
+            <SvgIcon size="mini" className={styles.claimCta} name={claimedData.logo} />
+          </TooltipTrigger>
+        )}
         content={guild}
         type="guilds"
         tabs={[
           {
-            name: 'Users',
+            name: 'Overview',
             to: `/g/${encodedGuildName}`,
             ignoreTitle: true,
+            content: (
+              <Overview data={guild} />
+            ),
+          },
+          {
+            name: 'Users',
+            to: `/g/${encodedGuildName}/users`,
             content: (
               <ContentCardList
                 noBorder